refactoring per evitare ciclo dipendenze
