+++ ABLV_IngSW/src/main/java/controller/svolgimentoGara/InizioGaraController.java
@@ -15,10 +15,10 @@ import javafx.stage.Stage;
 import javafx.util.StringConverter;
 import model.Concorrente;
 import model.Gara;
+import model.RisultatoTurno;
 import model.Settore;
 import service.PunteggioService;
 import service.RicercaService;
-import service.RisultatoTurno;
 import service.ArbitroService;
 import state.Session;
 import utils.Alerter;
