+++ ABLV_IngSW/src/main/java/model/RisultatoTurno.java
@@ -1,4 +1,4 @@
-package service;
+package model;
 
 public class RisultatoTurno {
